Added function weakFind to lookup a mapped value in the ConcurrentInsertOnlyHashMap.
The function weakFind returns the value associated with a key and provides a bit more
use than just weakContains, which is now expressed in terms of weakFind.

The weakFind function allows us to implement a fast path for implementing a cache, because
we can lookup values without requiring any kind of memory allocation, which the get() function
requires.