Fix incorrect subprocess exit handling.
First, the exit code produced by waitpid contains the priocess'
Second, exit(x) drops all but the least-significant 8 bits.
exit code in bits 8-15.
If the exit status from waitpid() is passed to exit(), then any error status
of the original process is dropped.

This commit fixes that and return EXIT_FAILURE if a signal
or anything else happened.